Refactor: 패키지 변경
- 패키지의 챕터 재분류
