+++ src/fastcampus/part2/chapter4/tree/bj15681/Main.java
@@ -1,4 +1,4 @@
-package fastcampus.part2.chapter3.tree.bj15681;
+package fastcampus.part2.chapter4.tree.bj15681;
 
 import java.io.BufferedReader;
 import java.io.IOException;
